requirements.txt
appdirs
click
python-dotenv
python-json-logger
-e .
